Added User class and I am able to send the user info to different activities.
Added User class to store the userID and the current authorization token. The userID may not be needed. The password may have to be sent in a different and most secure method to the database. Next step is to create the UI of the main menu page and the "Add new receipt" page and set up the REST API to handle POST requests.